requirements.txt
streamlit 
requests 
python-dotenv 
youtube-transcript-api 
psycopg2 
nltk 
openai 
crewai 
together 
openai
cohere 
supabase 
pgvector
tiktoken